\name{lab7sarasara-package}
\alias{lab7sarasara-package}
\alias{lab7sarasara}
\docType{package}
\title{
\packageTitle{lab7sarasara}
}
\description{
\packageDescription{lab7sarasara}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab7sarasara}
\packageIndices{lab7sarasara}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{lab7sarasara}

Maintainer: \packageMaintainer{lab7sarasara}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
