% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab7sarasara.R
\docType{package}
\name{lab7sarasara}
\alias{lab7sarasara}
\alias{lab7sarasara-package}
\title{A package for ridge regression model fitting}
\description{
A package to fit a ridge regression model
to a data set after a formula specified by the user. 
A ridge constant, chosen by the user, is used.
}
\author{
Sara Jesperson and Sara Johansson
}
