% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Reference class for ridge regression}
\description{
Create a ridgereg class containing all necessary methods for fitting a ridge regression model
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula for the ridge regression}

\item{\code{data}}{A data set for the ridge regression}

\item{\code{lambda}}{The ridge constant, a numeric scalar}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Prints out the estimated coefficients}

\item{\code{predict(newdata = NULL)}}{Prints out the predicted values or if newdata is used, prints out predicted values for the new data set}

\item{\code{print()}}{Prints out the formula and the estimated coefficients}

\item{\code{statistics()}}{Fits a ridge regression model}
}}

