% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize airport arrival delays}
\usage{
visualize_airport_delays()
}
\value{
A scatterplot where the size of the points indicates the arrival delay
}
\description{
A function to visualize the mean of airport arrival delays. Each airport is represented 
by their longitude and latitude.
}
